styles tweaking, bring out shared value props components for use in how to use preview course super tw
