+++ tutor/resources/styles/components/tours/value-prop.less
@@ -13,16 +13,19 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding-bottom: @spacing;
+  padding: @spacing @spacing*0.5;
+  color: @openstax-black;
+
   .beta {
     font-style: italic;
     font-size: 80%;
   }
   .heading {
-    margin: @spacing;
+    margin: 0 0 @spacing*0.25 0;
     padding: 0;
     text-align: center;
     font-size: 48px;
+    color: @openstax-tertiary;
   }
   .sub-heading {
     font-size: 18px;
@@ -34,36 +37,56 @@
   .column-content {
     display: flex;
     flex-direction: row;
-    padding: @spacing;
+    padding: @spacing 0;
     .column {
       background-repeat: no-repeat;
-      padding-top: @spacing*2;
-      background-size: 75px auto;
+      padding-top: @spacing*2.5;
+      background-size: auto 75px;
       background-position: center top;
+      text-align: center;
       flex: 1;
+
       + .column {
-        padding-left: @spacing;
+        margin-left: @spacing;
       }
+
       h3 {
         font-size: 16px;
         margin: 0 0 @spacing*0.25 0;
         font-weight: 700;
       }
+
+      p {
+        line-height: 1.4em;
+      }
     }
   }
   .btn {
     color: @tutor-neutral-lightest;
+    font-weight: 700;
     background-color: @tutor-tertiary-light;
     border-radius: 5px;
+    padding: @spacing*0.25 @spacing;
   }
 
   &.cc-to-tutor,
   &.welcome-to-tutor {
     .tutor-background-image("value-prop/tutor-background.svg");
+    background-size: cover;
+    background-position: bottom center;
+
     .column {
       &.spaced       { .tutor-background-image("value-prop/spaced-practice.svg"); }
-      &.personalized { .tutor-background-image("value-prop/personalized.svg");    }
-      &.low-cost     { .tutor-background-image("value-prop/low-cost.svg");        }
+      &.personalized {
+        .tutor-background-image("value-prop/personalized.svg");
+        background-size: 120px auto;
+        background-position: center 5px;
+      }
+      &.low-cost     {
+        .tutor-background-image("value-prop/low-cost.svg");
+        background-size: 100px auto;
+        background-position: center 15px;
+      }
       &.two-step     { .tutor-background-image("value-prop/two-step.svg");        }
     }
   }
@@ -72,6 +95,16 @@
       &.thanks           { .tutor-background-image("cc-sunset/thanks.svg");             }
       &.export-by        { .tutor-background-image("cc-sunset/export-by.svg");          }
       &.not-launched-yet { .tutor-background-image("cc-sunset/under-construction.svg"); }
+    }
+  }
+  &.course-preview {
+    .column {
+      background-size: auto 80px;
+
+      &.all-features              { .tutor-background-image("course-preview/all-features.svg");             }
+      &.view-analytics            { .tutor-background-image("course-preview/view-analytics.svg");           }
+      &.cant-save-work            { .tutor-background-image("course-preview/cant-save-work.svg");           }
+      &.view-textbook-questions   { .tutor-background-image("course-preview/view-textbook-questions.svg");  }
     }
   }
 }
